#1187 Fetch Aurora StockDeliveryItems

+++ database/migrations/2022_10_25_134349_create_supplier_products_table.php
@@ -31,8 +31,7 @@ return new class () extends Migration {
             $table->foreign('supplier_id')->references('id')->on('suppliers');
             $table->unsignedSmallInteger('agent_id')->nullable()->index();
             $table->foreign('agent_id')->references('id')->on('agents');
-
-            $table->unsignedSmallInteger('stock_id')->nullable()->nullable();
+            $table->unsignedInteger('stock_id')->nullable()->nullable();
             $table->foreign('stock_id')->references('id')->on('stocks');
             $table->string('state')->index()->default(SupplierProductStateEnum::IN_PROCESS->value);
             $table->boolean('is_available')->index()->default(true);
